Add files via upload
My team and I completed an internship at "TESSOLVE SEMICONDUCTOR PRIVATE LIMITED" where we focused on "Embedded & IoT programming." During this internship, we successfully executed a project titled "LPG CYLINDER WEIGHT AND GAS LEAKAGE MONITORING SYSTEM." Our primary objective was to raise awareness about the dangers of gas leaks, which sadly lead to the loss of innocent lives and property damage in many instances today. This application we developed has the potential to be highly beneficial for various settings, including companies and households, ultimately contributing to saving lives.

To detect gas leaks, we employed the MQ6 LPG gas sensor. When the sensor detects gas, it triggers the automatic closure of the gas knob using a servo motor. In addition to detecting leaks, we integrated a load cell to measure the weight of the gas cylinder. To alert individuals in the event of a gas leak, we included a buzzer that activates when the gas concentration exceeds safety limits. Furthermore, we integrated an LED indicator that lights up when the gas cylinder is approaching empty. 

Additionally, we implemented a user-friendly interface using an LCD display, enabling users to view real-time measurements. In summary, our project presents a comprehensive gas leakage alert system designed to detect gas leaks, monitor gas cylinder weight, and issue timely warnings to ensure the safety of those in the vicinity.